URL shortener service code

+++ src/main/java/com/io/infracloud/urlshortener/entity/Domain.java
@@ -1,6 +1,5 @@
 package com.io.infracloud.urlshortener.entity;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Column;
@@ -10,7 +9,6 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
-import lombok.Data;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
@@ -25,7 +23,7 @@ public class Domain {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "id")
-  private BigInteger id;
+  private Long id;
 
   @Column(name = "domain_name")
   private String domainName;
